abilities and turn structure both use flow builder
