+++ lib/realms/flows.rb
@@ -1,0 +1,6 @@
+require "realms/flows/builder"
+
+module Realms
+  module Flows
+  end
+end
