+++ lib/realms/flows/declarations.rb
@@ -1,0 +1,10 @@
+require "realms/flows/declarations/sequence"
+require "realms/flows/declarations/modal"
+require "realms/flows/declarations/loop"
+
+module Realms
+  module Flows
+    module Declarations
+    end
+  end
+end
