+++ lib/realms/flows/declarations/modal.rb
@@ -1,7 +1,7 @@
 module Realms
-  module Abilities
+  module Flows
     module Declarations
-      class Sequence
+      class Modal
         attr_reader :declarations
 
         def initialize(declarations:)
@@ -12,14 +12,18 @@ module Realms
           Evaluated.new(self, context)
         end
 
+        delegate :game, to: :context
+
         def auto?
-          declarations.all?(&:auto?)
+          false
         end
 
         class Evaluated
           attr_reader :declaration, :context
 
           delegate :game, to: :context
+          delegate :choose, to: :game
+          delegate :active_turn, :active_player, :passive_player, to: :game
 
           def initialize(declaration, context)
             @declaration = declaration
@@ -27,7 +31,7 @@ module Realms
           end
 
           def execute
-            declaration.declarations.each do |declaration|
+            choose(declaration.declarations, subject: context.card.name) do |declaration|
               effect = declaration.evaluate(context) 
               game.perform(effect)
             end
