--- lib/realms/phases.rb
@@ -1,4 +1,0 @@
-module Realms
-  module Phases
-  end
-end
