--- lib/realms/turns/declarations.rb
@@ -1,9 +1,0 @@
-require "realms/turns/declarations/phase"
-require "realms/turns/declarations/loop"
-
-module Realms
-  module Turns
-    module Declarations
-    end
-  end
-end
